move config_to_string to ocs_ci/utility/utils.py
