+++ ocs_ci/utility/workloads/fio.py
@@ -87,31 +87,3 @@ def run(**kwargs):
     log.info(f"Running cmd: {fio_cmd}")
 
     return io_pod.exec_cmd_on_pod(fio_cmd, out_yaml_format=False)
-
-
-def config_to_string(config):
-    """
-    Convert ConfigParser object to string in INI format.
-
-    Args:
-        config (obj): ConfigParser object
-
-    Returns:
-        str: Config in one string
-
-    """
-    string = ""
-    sections = [config.default_section] + config.sections()
-
-    first = True
-    for section in sections:
-        if config.items(section):
-            if first:
-                first = False
-            else:
-                string += "\n\n"
-            string += f"[{section}]\n"
-            string += "\n".join(
-                [f"{item[0]}={item[1]}" for item in config.items(section)]
-            )
-    return string
